Fix uninitialized variables
We can't really insert a phi function after an if statement because it
messes up uninitialized variable checking and regular phi function
placement.

Instead, we allocate an extra temp variable, then assign the return
value of each branch of the if statement to the temp value.  We'll use
normal phi placement to deal with the temp value later